Refactoring to other modules (part 1)
Left off at sacn_receiver_reset_networking
