+++ src/sacn/receiver.c
@@ -19,143 +19,13 @@
 
 #include "sacn/receiver.h"
 
-#include <limits.h>
 #include <stdint.h>
-#include <string.h>
-#include "etcpal/acn_pdu.h"
-#include "etcpal/acn_rlp.h"
 #include "etcpal/common.h"
-#include "etcpal/rbtree.h"
 #include "sacn/private/common.h"
 #include "sacn/private/source_loss.h"
 #include "sacn/private/mem.h"
-#include "sacn/private/pdu.h"
 #include "sacn/private/receiver.h"
-#include "sacn/private/sockets.h"
-#include "sacn/private/util.h"
-
-#if SACN_DYNAMIC_MEM
-#include <stdlib.h>
-#else
-#include "etcpal/mempool.h"
-#endif
-
-/***************************** Private constants *****************************/
-
-static const EtcPalThreadParams kReceiverThreadParams = {SACN_RECEIVER_THREAD_PRIORITY, SACN_RECEIVER_THREAD_STACK,
-                                                         "sACN Receive Thread", NULL};
-
-#define SACN_PERIODIC_INTERVAL 120
-
-/****************************** Private macros *******************************/
-
-/* Macros for dynamic vs static allocation. Static allocation is done using etcpal_mempool. */
-#if SACN_DYNAMIC_MEM
-#define ALLOC_RECEIVER() malloc(sizeof(SacnReceiver))
-#define ALLOC_TRACKED_SOURCE() malloc(sizeof(SacnTrackedSource))
-#define FREE_RECEIVER(ptr)        \
-  do                              \
-  {                               \
-    if (ptr->netints.netints)     \
-    {                             \
-      free(ptr->netints.netints); \
-    }                             \
-    free(ptr);                    \
-  } while (0)
-#define FREE_TRACKED_SOURCE(ptr) free(ptr)
-#else
-#define ALLOC_RECEIVER() etcpal_mempool_alloc(sacnrecv_receivers)
-#define ALLOC_TRACKED_SOURCE() etcpal_mempool_alloc(sacnrecv_tracked_sources)
-#define FREE_RECEIVER(ptr) etcpal_mempool_free(sacnrecv_receivers, ptr)
-#define FREE_TRACKED_SOURCE(ptr) etcpal_mempool_free(sacnrecv_tracked_sources, ptr)
-#endif
-
-/****************************** Private types ********************************/
-
-typedef struct PeriodicCallbacks
-{
-  const SourcesLostNotification* sources_lost_arr;
-  size_t num_sources_lost;
-  const SamplingStartedNotification* sampling_started_arr;
-  size_t num_sampling_started;
-  const SamplingEndedNotification* sampling_ended_arr;
-  size_t num_sampling_ended;
-} PeriodicCallbacks;
-
-/**************************** Private variables ******************************/
-
-#if !SACN_DYNAMIC_MEM
-ETCPAL_MEMPOOL_DEFINE(sacnrecv_receivers, SacnReceiver, SACN_RECEIVER_MAX_UNIVERSES);
-ETCPAL_MEMPOOL_DEFINE(sacnrecv_tracked_sources, SacnTrackedSource, SACN_RECEIVER_TOTAL_MAX_SOURCES);
-ETCPAL_MEMPOOL_DEFINE(sacnrecv_rb_nodes, EtcPalRbNode, SACN_RECEIVER_MAX_RB_NODES);
-#endif
-
-static struct SacnRecvState
-{
-  uint32_t expired_wait;
-
-  IntHandleManager handle_mgr;
-  EtcPalRbTree receivers;
-  EtcPalRbTree receivers_by_universe;
-} receiver_state;
-
-/*********************** Private function prototypes *************************/
-
-// Receiver creation and destruction
-static etcpal_error_t validate_receiver_config(const SacnReceiverConfig* config);
-static etcpal_error_t create_new_receiver(const SacnReceiverConfig* config, SacnMcastInterface* netints,
-                                          size_t num_netints, SacnReceiver** new_receiver);
-static etcpal_error_t assign_receiver_to_thread(SacnReceiver* receiver);
-static etcpal_error_t insert_receiver_into_maps(SacnReceiver* receiver);
-static etcpal_error_t start_receiver_thread(SacnRecvThreadContext* recv_thread_context);
-
-static etcpal_error_t add_receiver_sockets(SacnReceiver* receiver);
-
-static etcpal_error_t reset_receiver_state(SacnReceiver* receiver);
-
-static void remove_receiver_from_thread(SacnReceiver* receiver, socket_close_behavior_t socket_close_behavior);
-static void remove_receiver_from_maps(SacnReceiver* receiver);
-
-static void remove_receiver_sockets(SacnReceiver* receiver, socket_close_behavior_t close_behavior);
-static void remove_all_receiver_sockets(socket_close_behavior_t close_behavior);
-
-static void sacn_receive_thread(void* arg);
-
-// Receiving incoming data
-static void handle_incoming(sacn_thread_id_t thread_id, const uint8_t* data, size_t datalen,
-                            const EtcPalSockAddr* from_addr);
-static void handle_sacn_data_packet(sacn_thread_id_t thread_id, const uint8_t* data, size_t datalen,
-                                    const EtcPalUuid* sender_cid, const EtcPalSockAddr* from_addr);
-static void process_null_start_code(const SacnReceiver* receiver, SacnTrackedSource* src,
-                                    SourcePapLostNotification* source_pap_lost, bool* notify);
-static void process_pap(const SacnReceiver* receiver, SacnTrackedSource* src, bool* notify);
-static void process_new_source_data(SacnReceiver* receiver, const EtcPalUuid* sender_cid, const SacnHeaderData* header,
-                                    uint8_t seq, SacnTrackedSource** new_source,
-                                    SourceLimitExceededNotification* source_limit_exceeded, bool* notify);
-static bool check_sequence(int8_t new_seq, int8_t old_seq);
-static void deliver_receive_callbacks(const EtcPalSockAddr* from_addr, const EtcPalUuid* sender_cid,
-                                      const SacnHeaderData* header,
-                                      SourceLimitExceededNotification* source_limit_exceeded,
-                                      SourcePapLostNotification* source_pap_lost,
-                                      UniverseDataNotification* universe_data);
-
-// Process periodic timeout functionality
-static void process_receivers(SacnRecvThreadContext* recv_thread_context);
-static void process_receiver_sources(sacn_thread_id_t thread_id, SacnReceiver* receiver,
-                                     SourcesLostNotification* sources_lost);
-static bool check_source_timeouts(SacnTrackedSource* src, SacnSourceStatusLists* status_lists);
-static void update_source_status(SacnTrackedSource* src, SacnSourceStatusLists* status_lists);
-static void deliver_periodic_callbacks(const PeriodicCallbacks* periodic_callbacks);
-
-// Tree node management
-static int tracked_source_compare(const EtcPalRbTree* tree, const void* value_a, const void* value_b);
-static int receiver_compare(const EtcPalRbTree* tree, const void* value_a, const void* value_b);
-static int receiver_compare_by_universe(const EtcPalRbTree* tree, const void* value_a, const void* value_b);
-static EtcPalRbNode* node_alloc(void);
-static void node_dealloc(EtcPalRbNode* node);
-static void source_tree_dealloc(const EtcPalRbTree* self, EtcPalRbNode* node);
-static void universe_tree_dealloc(const EtcPalRbTree* self, EtcPalRbNode* node);
-static bool receiver_handle_in_use(int handle_val, void* cookie);
+#include "sacn/private/receiver_state.h"
 
 /*************************** Function definitions ****************************/
 
@@ -166,48 +36,13 @@ static bool receiver_handle_in_use(int handle_val, void* cookie);
 /* Initialize the sACN Receiver module. Internal function called from sacn_init(). */
 etcpal_error_t sacn_receiver_init(void)
 {
-  etcpal_error_t res = kEtcPalErrOk;
-
-#if !SACN_DYNAMIC_MEM
-  res |= etcpal_mempool_init(sacnrecv_receivers);
-  res |= etcpal_mempool_init(sacnrecv_tracked_sources);
-  res |= etcpal_mempool_init(sacnrecv_rb_nodes);
-#endif
-
-  if (res == kEtcPalErrOk)
-  {
-    etcpal_rbtree_init(&receiver_state.receivers, receiver_compare, node_alloc, node_dealloc);
-    etcpal_rbtree_init(&receiver_state.receivers_by_universe, receiver_compare_by_universe, node_alloc, node_dealloc);
-    init_int_handle_manager(&receiver_state.handle_mgr, receiver_handle_in_use, NULL);
-    receiver_state.expired_wait = SACN_DEFAULT_EXPIRED_WAIT_MS;
-  }
-  else
-  {
-    memset(&receiver_state, 0, sizeof receiver_state);
-  }
-
-  return res;
+  return kEtcPalErrOk;  // Nothing to do here.
 }
 
 /* Deinitialize the sACN Receiver module. Internal function called from sacn_deinit(). */
 void sacn_receiver_deinit(void)
 {
-  // Stop all receive threads
-  for (unsigned int i = 0; i < sacn_mem_get_num_threads(); ++i)
-  {
-    SacnRecvThreadContext* thread_context = get_recv_thread_context(i);
-    if (thread_context && thread_context->running)
-    {
-      thread_context->running = false;
-      etcpal_thread_join(&thread_context->thread_handle);
-      sacn_cleanup_dead_sockets(thread_context);
-    }
-  }
-
-  // Clear out the rest of the state tracking
-  etcpal_rbtree_clear_with_cb(&receiver_state.receivers, universe_tree_dealloc);
-  etcpal_rbtree_clear(&receiver_state.receivers_by_universe);
-  memset(&receiver_state, 0, sizeof receiver_state);
+  // Nothing to do here.
 }
 
 /**
@@ -254,37 +89,33 @@ void sacn_receiver_config_init(SacnReceiverConfig* config)
 etcpal_error_t sacn_receiver_create(const SacnReceiverConfig* config, sacn_receiver_t* handle,
                                     SacnMcastInterface* netints, size_t num_netints)
 {
-  if (!config || !handle)
-    return kEtcPalErrInvalid;
+  etcpal_error_t res = kEtcPalErrOk;
 
   if (!sacn_initialized())
-    return kEtcPalErrNotInit;
-
-  etcpal_error_t res = validate_receiver_config(config);
-  if (res != kEtcPalErrOk)
-    return res;
+  {
+    res = kEtcPalErrNotInit;
+  }
+  else if (!config || !handle)
+  {
+    res = kEtcPalErrInvalid;
+  }
+  else if (!UNIVERSE_ID_VALID(config->universe_id) || !config->callbacks.universe_data ||
+           !config->callbacks.sources_lost || !config->callbacks.sampling_period_ended)
+  {
+    res = kEtcPalErrInvalid;
+  }
 
   if (sacn_lock())
   {
-    // First check to see if we are already listening on this universe.
-    SacnReceiverKeys lookup_keys;
-    lookup_keys.universe = config->universe_id;
-    SacnReceiver* receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers_by_universe, &lookup_keys);
-    if (receiver)
+    SacnReceiver* receiver = NULL;
+    if (res == kEtcPalErrOk)
+      res = add_sacn_receiver(get_next_receiver_handle(), config, netints, num_netints, &receiver);
+
+    if (res == kEtcPalErrOk)
     {
-      res = kEtcPalErrExists;
-      receiver = NULL;
-    }
-
-    if (res == kEtcPalErrOk)
-      res = create_new_receiver(config, netints, num_netints, &receiver);
-
-    if (res == kEtcPalErrOk)
+      begin_sampling_period(receiver);
       res = assign_receiver_to_thread(receiver);
-
-    // Insert the new universe into the map.
-    if (res == kEtcPalErrOk)
-      res = insert_receiver_into_maps(receiver);
+    }
 
     if (res == kEtcPalErrOk)
     {
@@ -294,9 +125,8 @@ etcpal_error_t sacn_receiver_create(const SacnReceiverConfig* config, sacn_recei
     {
       if (receiver)
       {
-        remove_receiver_from_maps(receiver);
         remove_receiver_from_thread(receiver, kCloseSocketNow);
-        FREE_RECEIVER(receiver);
+        remove_sacn_receiver(receiver);
       }
     }
     sacn_unlock();
@@ -323,30 +153,30 @@ etcpal_error_t sacn_receiver_create(const SacnReceiverConfig* config, sacn_recei
  */
 etcpal_error_t sacn_receiver_destroy(sacn_receiver_t handle)
 {
-  if (!sacn_initialized())
-    return kEtcPalErrNotInit;
-
   etcpal_error_t res = kEtcPalErrOk;
+
+  if (!sacn_initialized())
+    res = kEtcPalErrNotInit;
+
   if (sacn_lock())
   {
-    SacnReceiver* receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers, &handle);
-    if (receiver)
+    SacnReceiver* receiver = NULL;
+    if (res == kEtcPalErrOk)
+      res = lookup_receiver(handle, &receiver);
+
+    if (res == kEtcPalErrOk)
     {
       remove_receiver_from_thread(receiver, kQueueSocketForClose);
-      etcpal_rbtree_clear_with_cb(&receiver->sources, source_tree_dealloc);
-      remove_receiver_from_maps(receiver);
-      FREE_RECEIVER(receiver);
+      remove_sacn_receiver(receiver);
     }
-    else
-    {
-      res = kEtcPalErrNotFound;
-    }
+
     sacn_unlock();
   }
   else
   {
     res = kEtcPalErrSys;
   }
+
   return res;
 }
 
@@ -363,19 +193,21 @@ etcpal_error_t sacn_receiver_destroy(sacn_receiver_t handle)
  */
 etcpal_error_t sacn_receiver_get_universe(sacn_receiver_t handle, uint16_t* universe_id)
 {
-  if (!sacn_initialized())
-    return kEtcPalErrNotInit;
-  if (universe_id == NULL)
-    return kEtcPalErrInvalid;
-
   etcpal_error_t res = kEtcPalErrOk;
+
+  if (!sacn_initialized())
+    res = kEtcPalErrNotInit;
+  else if (universe_id == NULL)
+    res = kEtcPalErrInvalid;
+
   if (sacn_lock())
   {
-    SacnReceiver* receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers, &handle);
-    if (receiver)
+    SacnReceiver* receiver = NULL;
+    if (res == kEtcPalErrOk)
+      res = lookup_receiver(handle, &receiver);
+
+    if (res == kEtcPalErrOk)
       *universe_id = receiver->keys.universe;
-    else
-      res = kEtcPalErrNotFound;
 
     sacn_unlock();
   }
@@ -402,47 +234,35 @@ etcpal_error_t sacn_receiver_get_universe(sacn_receiver_t handle, uint16_t* univ
  */
 etcpal_error_t sacn_receiver_change_universe(sacn_receiver_t handle, uint16_t new_universe_id)
 {
-  if (!UNIVERSE_ID_VALID(new_universe_id))
-    return kEtcPalErrInvalid;
+  etcpal_error_t res = kEtcPalErrOk;
 
   if (!sacn_initialized())
-    return kEtcPalErrNotInit;
+    res = kEtcPalErrNotInit;
+  else if (!UNIVERSE_ID_VALID(new_universe_id))
+    res = kEtcPalErrInvalid;
 
-  etcpal_error_t res = kEtcPalErrOk;
   if (sacn_lock())
   {
     // First check to see if there is already a receiver listening on this universe.
-    SacnReceiverKeys lookup_keys;
-    lookup_keys.universe = new_universe_id;
-    SacnReceiver* receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers_by_universe, &lookup_keys);
-    if (receiver)
+    if (res == kEtcPalErrOk)
     {
-      res = kEtcPalErrExists;
+      SacnReceiver* tmp = NULL;
+      if (lookup_receiver_by_universe(new_universe_id, &tmp) == kEtcPalErrOk)
+        res = kEtcPalErrExists;
     }
 
     // Find the receiver to change the universe for.
+    SacnReceiver* receiver = NULL;
     if (res == kEtcPalErrOk)
-    {
-      receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers, &handle);
-      if (!receiver)
-        res = kEtcPalErrNotFound;
-    }
+      res = lookup_receiver(handle, &receiver);
 
     // Clear termination sets and sources since they only pertain to the old universe.
     if (res == kEtcPalErrOk)
-    {
-      clear_term_set_list(receiver->term_sets);
-      receiver->term_sets = NULL;
-      res = etcpal_rbtree_clear_with_cb(&receiver->sources, source_tree_dealloc);
-    }
+      res = clear_term_sets_and_sources(receiver);
 
     // Update receiver key and position in receiver_state.receivers_by_universe.
     if (res == kEtcPalErrOk)
-    {
-      etcpal_rbtree_remove(&receiver_state.receivers_by_universe, receiver);
-      receiver->keys.universe = new_universe_id;
-      res = etcpal_rbtree_insert(&receiver_state.receivers_by_universe, receiver);
-    }
+      res = update_receiver_universe(receiver, new_universe_id);
 
     // Update the receiver's socket and subscription.
     if (res == kEtcPalErrOk)
@@ -453,11 +273,7 @@ etcpal_error_t sacn_receiver_change_universe(sacn_receiver_t handle, uint16_t ne
 
     // Begin the sampling period.
     if (res == kEtcPalErrOk)
-    {
-      receiver->sampling = true;
-      receiver->notified_sampling_started = false;
-      etcpal_timer_start(&receiver->sample_timer, SAMPLE_TIME);
-    }
+      begin_sampling_period(receiver);
 
     sacn_unlock();
   }
@@ -688,1066 +504,4 @@ uint32_t sacn_receiver_get_expired_wait()
     sacn_unlock();
   }
   return res;
-}
-
-/**************************************************************************************************
- * Internal helpers for receiver creation and destruction
- *************************************************************************************************/
-
-/*
- * Make sure the values provided in an SacnReceiverConfig struct are valid.
- */
-etcpal_error_t validate_receiver_config(const SacnReceiverConfig* config)
-{
-  SACN_ASSERT(config);
-
-  if (!UNIVERSE_ID_VALID(config->universe_id) || !config->callbacks.universe_data || !config->callbacks.sources_lost ||
-      !config->callbacks.sampling_period_ended)
-  {
-    return kEtcPalErrInvalid;
-  }
-
-  return kEtcPalErrOk;
-}
-
-/*
- * Allocate a new receiver instances and do essential first initialization, in preparation for
- * creating the sockets and subscriptions.
- *
- * [in] config Receiver configuration data.
- * Returns the new initialized receiver instance, or NULL if out of memory.
- */
-etcpal_error_t create_new_receiver(const SacnReceiverConfig* config, SacnMcastInterface* netints, size_t num_netints,
-                                   SacnReceiver** new_receiver)
-{
-  SACN_ASSERT(config);
-
-  sacn_receiver_t new_handle = get_next_int_handle(&receiver_state.handle_mgr, -1);
-  if (new_handle == SACN_RECEIVER_INVALID)
-    return kEtcPalErrNoMem;
-
-  SacnReceiver* receiver = ALLOC_RECEIVER();
-  if (!receiver)
-    return kEtcPalErrNoMem;
-
-  receiver->keys.handle = new_handle;
-  receiver->keys.universe = config->universe_id;
-  receiver->thread_id = SACN_THREAD_ID_INVALID;
-
-  receiver->ipv4_socket = ETCPAL_SOCKET_INVALID;
-  receiver->ipv6_socket = ETCPAL_SOCKET_INVALID;
-
-  etcpal_error_t initialize_receiver_netints_result =
-      sacn_initialize_receiver_netints(&receiver->netints, netints, num_netints);
-  if (initialize_receiver_netints_result != kEtcPalErrOk)
-  {
-    FREE_RECEIVER(receiver);
-    return initialize_receiver_netints_result;
-  }
-
-  receiver->sampling = true;
-  receiver->notified_sampling_started = false;
-  etcpal_timer_start(&receiver->sample_timer, SAMPLE_TIME);
-  receiver->suppress_limit_exceeded_notification = false;
-  etcpal_rbtree_init(&receiver->sources, tracked_source_compare, node_alloc, node_dealloc);
-  receiver->term_sets = NULL;
-
-  receiver->filter_preview_data = ((config->flags & SACN_RECEIVER_OPTS_FILTER_PREVIEW_DATA) != 0);
-
-  receiver->callbacks = config->callbacks;
-
-  receiver->source_count_max = config->source_count_max;
-
-  receiver->ip_supported = config->ip_supported;
-
-  receiver->next = NULL;
-
-  *new_receiver = receiver;
-
-  return kEtcPalErrOk;
-}
-
-/*
- * Pick a thread for the receiver based on current load balancing, create the receiver's sockets,
- * and assign it to that thread.
- *
- * [in,out] receiver Receiver instance to assign. Make sure the universe ID is up-to-date.
- * Returns error code indicating the result of the operations.
- */
-etcpal_error_t assign_receiver_to_thread(SacnReceiver* receiver)
-{
-  SacnRecvThreadContext* assigned_thread = NULL;
-
-  // Assign this receiver to the thread with the lowest number of receivers currently
-  for (unsigned int i = 0; i < sacn_mem_get_num_threads(); ++i)
-  {
-    SacnRecvThreadContext* context = get_recv_thread_context(i);
-    if (context)
-    {
-      if (!assigned_thread || (context->num_receivers < assigned_thread->num_receivers))
-      {
-        receiver->thread_id = i;
-        assigned_thread = context;
-      }
-    }
-  }
-
-  SACN_ASSERT(assigned_thread);
-
-  etcpal_error_t res = add_receiver_sockets(receiver);
-
-  if ((res == kEtcPalErrOk) && !assigned_thread->running)
-  {
-    res = start_receiver_thread(assigned_thread);
-    if (res != kEtcPalErrOk)
-    {
-      remove_receiver_sockets(receiver, kCloseSocketNow);
-    }
-  }
-
-  if (res == kEtcPalErrOk)
-  {
-    // Append the receiver to the thread list
-    add_receiver_to_list(assigned_thread, receiver);
-  }
-
-  return res;
-}
-
-/*
- * Add a receiver to the maps that are used to track receivers globally.
- *
- * [in] receiver Receiver instance to add.
- * Returns error code indicating the result of the operations.
- */
-etcpal_error_t insert_receiver_into_maps(SacnReceiver* receiver)
-{
-  etcpal_error_t res = etcpal_rbtree_insert(&receiver_state.receivers, receiver);
-  if (res == kEtcPalErrOk)
-  {
-    res = etcpal_rbtree_insert(&receiver_state.receivers_by_universe, receiver);
-    if (res != kEtcPalErrOk)
-      etcpal_rbtree_remove(&receiver_state.receivers, receiver);
-  }
-  return res;
-}
-
-/*
- * Start a new thread to process receiver state. The thread is associated with a specific
- * SacnRecvThreadContext instance.
- *
- * [in,out] recv_thread_context Thread context corresponding to the thread to start.
- * Returns error code indicating the result of the thread start operation.
- */
-etcpal_error_t start_receiver_thread(SacnRecvThreadContext* recv_thread_context)
-{
-  recv_thread_context->running = true;
-  etcpal_error_t create_res = etcpal_thread_create(&recv_thread_context->thread_handle, &kReceiverThreadParams,
-                                                   sacn_receive_thread, recv_thread_context);
-  if (create_res != kEtcPalErrOk)
-  {
-    recv_thread_context->running = false;
-  }
-  return create_res;
-}
-
-/*
- * Initialize a receiver's IPv4 and IPv6 sockets. Make sure to take the sACN lock before calling.
- *
- * [in] receiver Receiver to add sockets for. Make sure the universe ID, thread ID, IP supported, and netints are
- * initialized.
- * Returns error code indicating the result of adding the sockets.
- */
-etcpal_error_t add_receiver_sockets(SacnReceiver* receiver)
-{
-  etcpal_error_t res = kEtcPalErrOk;
-
-  if (supports_ipv4(receiver->ip_supported))
-  {
-    res = sacn_add_receiver_socket(receiver->thread_id, kEtcPalIpTypeV4, receiver->keys.universe,
-                                   receiver->netints.netints, receiver->netints.num_netints, &receiver->ipv4_socket);
-
-    if ((res == kEtcPalErrNoNetints) && supports_ipv6(receiver->ip_supported))
-      res = kEtcPalErrOk;  // Try IPv6.
-  }
-
-  if ((res == kEtcPalErrOk) && supports_ipv6(receiver->ip_supported))
-  {
-    res = sacn_add_receiver_socket(receiver->thread_id, kEtcPalIpTypeV6, receiver->keys.universe,
-                                   receiver->netints.netints, receiver->netints.num_netints, &receiver->ipv6_socket);
-  }
-
-  return res;
-}
-
-/*
- * Reset a receiver's state. Make sure to take the sACN lock before calling.
- *
- * [in,out] receiver Receiver to reset state for.
- * Returns error code indicating the result of the reset.
- */
-etcpal_error_t reset_receiver_state(SacnReceiver* receiver)
-{
-  receiver->sampling = true;
-  receiver->notified_sampling_started = false;
-  etcpal_timer_start(&receiver->sample_timer, SAMPLE_TIME);
-  receiver->suppress_limit_exceeded_notification = false;
-
-  clear_term_set_list(receiver->term_sets);
-  receiver->term_sets = NULL;
-
-  return etcpal_rbtree_clear_with_cb(&receiver->sources, source_tree_dealloc);
-}
-
-/*
- * Remove a receiver instance from a receiver thread. After this completes, the thread will no
- * longer process timeouts for that receiver.
- *
- * [in,out] receiver Receiver to remove.
- * [in] socket_close_behavior Whether to close the socket immediately (e.g. on full library shutdown).
- * Returns error code indicating the result of the removal operation.
- */
-void remove_receiver_from_thread(SacnReceiver* receiver, socket_close_behavior_t socket_close_behavior)
-{
-  SacnRecvThreadContext* context = get_recv_thread_context(receiver->thread_id);
-  if (context)
-  {
-    remove_receiver_sockets(receiver, socket_close_behavior);
-
-    remove_receiver_from_list(context, receiver);
-  }
-}
-
-/*
- * Remove a receiver instance from the maps that are used to track receivers globally.
- *
- * [in] receiver Receiver to remove.
- */
-void remove_receiver_from_maps(SacnReceiver* receiver)
-{
-  etcpal_rbtree_remove(&receiver_state.receivers_by_universe, receiver);
-  etcpal_rbtree_remove(&receiver_state.receivers, receiver);
-}
-
-/*
- * Remove a receiver's sockets, choosing whether to close them now or wait until the next thread cycle.
- *
- * [in/out] receiver Receiver whose sockets to remove. Socket handles are set to invalid.
- * [in] close_behavior Whether to close the sockets now or wait until the next thread cycle.
- */
-void remove_receiver_sockets(SacnReceiver* receiver, socket_close_behavior_t close_behavior)
-{
-  if (receiver->ipv4_socket != ETCPAL_SOCKET_INVALID)
-    sacn_remove_receiver_socket(receiver->thread_id, &receiver->ipv4_socket, close_behavior);
-  if (receiver->ipv6_socket != ETCPAL_SOCKET_INVALID)
-    sacn_remove_receiver_socket(receiver->thread_id, &receiver->ipv6_socket, close_behavior);
-}
-
-/*
- * Remove all receiver sockets, choosing whether to close them now or wait until the next thread cycle.
- *
- * The sACN lock should be locked before calling this.
- *
- * [in] close_behavior Whether to close the sockets now or wait until the next thread cycle.
- */
-void remove_all_receiver_sockets(socket_close_behavior_t close_behavior)
-{
-  EtcPalRbIter iter;
-  etcpal_rbiter_init(&iter);
-  for (SacnReceiver* receiver = etcpal_rbiter_first(&iter, &receiver_state.receivers); receiver;
-       receiver = etcpal_rbiter_next(&iter))
-  {
-    remove_receiver_sockets(receiver, close_behavior);
-  }
-}
-
-/*
- * The receiver thread function. Receives and forwards sACN data from the network and processes
- * periodic timeouts for sACN receivers.
- */
-void sacn_receive_thread(void* arg)
-{
-  SacnRecvThreadContext* context = (SacnRecvThreadContext*)arg;
-  SACN_ASSERT(context);
-
-  // Create the poll context
-  etcpal_error_t init_res = etcpal_poll_context_init(&context->poll_context);
-  if (init_res != kEtcPalErrOk)
-  {
-    SACN_LOG_CRIT(
-        "Could not create a socket poll context for sACN: '%s'. sACN Receive functionality will not work properly.",
-        etcpal_strerror(init_res));
-    return;
-  }
-
-  EtcPalTimer periodic_timer;
-  etcpal_timer_start(&periodic_timer, SACN_PERIODIC_INTERVAL);
-
-  while (context->running)
-  {
-    if (sacn_lock())
-    {
-      sacn_add_pending_sockets(context);
-      sacn_cleanup_dead_sockets(context);
-      sacn_unlock();
-    }
-
-    SacnReadResult read_result;
-    etcpal_error_t read_res = sacn_read(context, &read_result);
-    if (read_res == kEtcPalErrOk)
-    {
-      handle_incoming(context->thread_id, read_result.data, read_result.data_len, &read_result.from_addr);
-    }
-    else if (read_res != kEtcPalErrTimedOut)
-    {
-      if (read_res != kEtcPalErrNoSockets)
-      {
-        SACN_LOG_WARNING("Error occurred while attempting to read sACN incoming data: '%s'.",
-                         etcpal_strerror(read_res));
-      }
-      etcpal_thread_sleep(SACN_RECEIVER_READ_TIMEOUT_MS);
-    }
-
-    if (etcpal_timer_is_expired(&periodic_timer))
-    {
-      process_receivers(context);
-      etcpal_timer_reset(&periodic_timer);
-    }
-  }
-
-  // Destroy the poll context
-  etcpal_poll_context_deinit(&context->poll_context);
-}
-
-/**************************************************************************************************
- * Internal helpers for processing incoming sACN data
- *************************************************************************************************/
-
-/*
- * Handle an incoming data packet on a receiver socket.
- *
- * [in] thread_id ID for the thread in which the incoming data was received.
- * [in] data Incoming data buffer.
- * [in] datalen Size of data buffer.
- * [in] from_addr Network address from which the data was received.
- */
-void handle_incoming(sacn_thread_id_t thread_id, const uint8_t* data, size_t datalen, const EtcPalSockAddr* from_addr)
-{
-  AcnUdpPreamble preamble;
-  if (!acn_parse_udp_preamble(data, datalen, &preamble))
-    return;
-
-  AcnRootLayerPdu rlp;
-  AcnPdu lpdu = ACN_PDU_INIT;
-  while (acn_parse_root_layer_pdu(preamble.rlp_block, preamble.rlp_block_len, &rlp, &lpdu))
-  {
-    if (rlp.vector == ACN_VECTOR_ROOT_E131_DATA)
-      handle_sacn_data_packet(thread_id, rlp.pdata, rlp.data_len, &rlp.sender_cid, from_addr);
-  }
-}
-
-/*
- * Handle an sACN Data packet that has been unpacked from a Root Layer PDU.
- *
- * [in] thread_id ID for the thread in which the data packet was received.
- * [in] data Buffer containing the data packet.
- * [in] datalen Size of buffer.
- * [in] sender_cid CID from which the data was received.
- * [in] from_addr Network address from which the data was received.
- */
-void handle_sacn_data_packet(sacn_thread_id_t thread_id, const uint8_t* data, size_t datalen,
-                             const EtcPalUuid* sender_cid, const EtcPalSockAddr* from_addr)
-{
-  UniverseDataNotification* universe_data = get_universe_data(thread_id);
-  SourceLimitExceededNotification* source_limit_exceeded = get_source_limit_exceeded(thread_id);
-  SourcePapLostNotification* source_pap_lost = get_source_pap_lost(thread_id);
-  if (!universe_data || !source_limit_exceeded || !source_pap_lost)
-  {
-    SACN_LOG_ERR("Could not allocate memory for incoming sACN data packet!");
-    return;
-  }
-
-  uint8_t seq;
-  bool is_termination_packet;
-  bool parse_res = false;
-
-  SacnHeaderData* header = &universe_data->header;
-  header->cid = *sender_cid;
-
-  parse_res = parse_sacn_data_packet(data, datalen, header, &seq, &is_termination_packet, &universe_data->pdata);
-
-  if (!parse_res)
-  {
-    if (SACN_CAN_LOG(ETCPAL_LOG_WARNING))
-    {
-      char cid_str[ETCPAL_UUID_STRING_BYTES];
-      etcpal_uuid_to_string(sender_cid, cid_str);
-      SACN_LOG_WARNING("Ignoring malformed sACN data packet from component %s", cid_str);
-    }
-    return;
-  }
-
-  // Ignore SACN_STARTCODE_PRIORITY packets if SACN_ETC_PRIORITY_EXTENSION is disabled.
-#if !SACN_ETC_PRIORITY_EXTENSION
-  if (header->start_code == SACN_STARTCODE_PRIORITY)
-    return;
-#endif
-
-  if (sacn_lock())
-  {
-    SacnReceiverKeys lookup_keys;
-    lookup_keys.universe = header->universe_id;
-    SacnReceiver* receiver = (SacnReceiver*)etcpal_rbtree_find(&receiver_state.receivers_by_universe, &lookup_keys);
-    if (!receiver)
-    {
-      // We are not listening to this universe.
-      sacn_unlock();
-      return;
-    }
-
-    bool notify = false;
-    SacnTrackedSource* src = (SacnTrackedSource*)etcpal_rbtree_find(&receiver->sources, sender_cid);
-    if (src)
-    {
-      // Check to see if the 'stream terminated' bit is set in the options
-      if (is_termination_packet)
-      {
-        src->terminated = true;
-        etcpal_timer_start(&src->packet_timer, 0);
-      }
-      // This also handles the case where the source was already terminated in a previous packet
-      // but not yet removed.
-      if (src->terminated)
-      {
-        sacn_unlock();
-        return;
-      }
-
-      if (!check_sequence(seq, src->seq))
-      {
-        // Drop the packet
-        sacn_unlock();
-        return;
-      }
-      src->seq = seq;
-
-      // Based on the start code, update the timers.
-      if (header->start_code == SACN_STARTCODE_DMX)
-      {
-        process_null_start_code(receiver, src, source_pap_lost, &notify);
-      }
-#if SACN_ETC_PRIORITY_EXTENSION
-      else if (header->start_code == SACN_STARTCODE_PRIORITY)
-      {
-        process_pap(receiver, src, &notify);
-      }
-#endif
-      else if (header->start_code != SACN_STARTCODE_PRIORITY)
-      {
-        notify = true;
-      }
-    }
-    else if (!is_termination_packet)
-    {
-      process_new_source_data(receiver, sender_cid, header, seq, &src, source_limit_exceeded, &notify);
-    }
-    // Else we weren't tracking this source before and it is a termination packet. Ignore.
-
-    if (src)
-    {
-      if (header->preview && receiver->filter_preview_data)
-      {
-        notify = false;
-      }
-
-      if (notify)
-      {
-        universe_data->callback = receiver->callbacks.universe_data;
-        universe_data->handle = receiver->keys.handle;
-        universe_data->universe = receiver->keys.universe;
-        universe_data->is_sampling = receiver->sampling;
-        universe_data->context = receiver->callbacks.context;
-      }
-    }
-
-    sacn_unlock();
-  }
-
-  // Deliver callbacks if applicable.
-  deliver_receive_callbacks(from_addr, sender_cid, header, source_limit_exceeded, source_pap_lost, universe_data);
-}
-
-/*
- * Process the timers and logic upon receiving NULL START Code data from an existing source.
- *
- * [in] receiver Receiver on which this source data was received.
- * [in,out] src Existing source from which this data was received - state tracking is updated.
- * [out] source_pap_lost Notification data to deliver if a PAP lost condition should be forwarded
- *                       to the app.
- * [out] notify Whether or not to forward the data to the user in a notification.
- */
-void process_null_start_code(const SacnReceiver* receiver, SacnTrackedSource* src,
-                             SourcePapLostNotification* source_pap_lost, bool* notify)
-{
-  *notify = true;  // Notify universe data during and after the sampling period.
-
-  // No matter how valid, we got something.
-  src->dmx_received_since_last_tick = true;
-  etcpal_timer_start(&src->packet_timer, SOURCE_LOSS_TIMEOUT);
-
-#if SACN_ETC_PRIORITY_EXTENSION
-  switch (src->recv_state)
-  {
-    case kRecvStateWaitingForDmx:
-      // We had previously received PAP, were waiting for DMX and got it.
-      if (receiver->sampling)
-      {
-        // We are in the sample period - notify immediately.
-        src->recv_state = kRecvStateHaveDmxAndPap;
-      }
-      else
-      {
-        // Now we wait for one more PAP packet before notifying.
-        src->recv_state = kRecvStateWaitingForPap;
-        *notify = false;
-      }
-      break;
-    case kRecvStateWaitingForPap:
-      if (etcpal_timer_is_expired(&src->pap_timer))
-      {
-        // Our per-address-priority waiting period has expired. Keep the timer going in case the
-        // source starts sending PAP later.
-        src->recv_state = kRecvStateHaveDmxOnly;
-        etcpal_timer_start(&src->pap_timer, SOURCE_LOSS_TIMEOUT);
-      }
-      else
-      {
-        // We've received a DMX packet during our per-address-priority waiting period. Don't notify.
-        *notify = false;
-      }
-      break;
-    case kRecvStateHaveDmxOnly:
-      // More DMX, nothing to see here
-      break;
-    case kRecvStateHaveDmxAndPap:
-      if (etcpal_timer_is_expired(&src->pap_timer))
-      {
-        // Source stopped sending PAP but is still sending DMX.
-        // In this case, also notify the source_pap_lost callback.
-        source_pap_lost->callback = receiver->callbacks.source_pap_lost;
-        source_pap_lost->source.cid = src->cid;
-        ETCPAL_MSVC_NO_DEP_WRN strcpy(source_pap_lost->source.name, src->name);
-        source_pap_lost->context = receiver->callbacks.context;
-        source_pap_lost->handle = receiver->keys.handle;
-        source_pap_lost->universe = receiver->keys.universe;
-
-        src->recv_state = kRecvStateHaveDmxOnly;
-      }
-      break;
-    default:
-      break;
-  }
-#endif
-}
-
-#if SACN_ETC_PRIORITY_EXTENSION
-/*
- * Process the timers and logic upon receiving per-address priority data from an existing source.
- *
- * [in] receiver Receiver instance for which PAP data was received.
- * [in,out] src Existing source from which this PAP data was received - state tracking is updated.
- * [out] notify Whether or not to forward the data to the user in a notification.
- */
-void process_pap(const SacnReceiver* receiver, SacnTrackedSource* src, bool* notify)
-{
-  ETCPAL_UNUSED_ARG(receiver);
-
-  *notify = true;
-
-  switch (src->recv_state)
-  {
-    case kRecvStateWaitingForDmx:
-      // Still waiting for DMX - ignore PAP packets until we've seen at least one DMX packet.
-      *notify = false;
-      etcpal_timer_reset(&src->pap_timer);
-      break;
-    case kRecvStateWaitingForPap:
-    case kRecvStateHaveDmxOnly:
-      src->recv_state = kRecvStateHaveDmxAndPap;
-      etcpal_timer_start(&src->pap_timer, SOURCE_LOSS_TIMEOUT);
-      break;
-    case kRecvStateHaveDmxAndPap:
-      etcpal_timer_reset(&src->pap_timer);
-      break;
-    default:
-      break;
-  }
-}
-#endif
-
-/*
- * Process the timers and logic upon receiving data from a source we are not tracking yet.
- *
- * [in,out] receiver Receiver for which this source data was received - new source is added to its
- *                   tree.
- * [in] sender_cid CID of the sACN source that sent this data.
- * [in] header Header data contained in the sACN packet.
- * [in] seq Sequence number of the sACN packet.
- * [out] source_limit_exceeded Notification data to deliver if a source limit exceeded
- * condition should be forwarded to the app.
- * [out] notify Whether or not to forward the data to the user in a notification.
- */
-void process_new_source_data(SacnReceiver* receiver, const EtcPalUuid* sender_cid, const SacnHeaderData* header,
-                             uint8_t seq, SacnTrackedSource** new_source,
-                             SourceLimitExceededNotification* source_limit_exceeded, bool* notify)
-{
-#if SACN_ETC_PRIORITY_EXTENSION
-  if ((header->start_code != SACN_STARTCODE_DMX) && (header->start_code != SACN_STARTCODE_PRIORITY))
-    return;
-#else
-  if (header->start_code != SACN_STARTCODE_DMX)
-    return;
-#endif
-
-  // Notify universe data during and after the sampling period.
-  *notify = true;
-
-  // A new source has appeared!
-  SacnTrackedSource* src = NULL;
-
-  size_t current_number_of_sources = etcpal_rbtree_size(&receiver->sources);
-#if SACN_DYNAMIC_MEM
-  size_t max_number_of_sources = receiver->source_count_max;
-  bool infinite_sources = (max_number_of_sources == SACN_RECEIVER_INFINITE_SOURCES);
-#else
-  size_t max_number_of_sources = SACN_RECEIVER_MAX_SOURCES_PER_UNIVERSE;
-  bool infinite_sources = false;
-#endif
-
-  if (infinite_sources || (current_number_of_sources < max_number_of_sources))
-    src = ALLOC_TRACKED_SOURCE();
-
-  if (!src)
-  {
-    // No room for new source.
-    if (!receiver->suppress_limit_exceeded_notification)
-    {
-      receiver->suppress_limit_exceeded_notification = true;
-      source_limit_exceeded->callback = receiver->callbacks.source_limit_exceeded;
-      source_limit_exceeded->context = receiver->callbacks.context;
-      source_limit_exceeded->handle = receiver->keys.handle;
-      source_limit_exceeded->universe = receiver->keys.universe;
-      return;
-    }
-    else
-    {
-      // Notification suppressed - don't notify
-      return;
-    }
-  }
-  src->cid = *sender_cid;
-  ETCPAL_MSVC_NO_DEP_WRN strcpy(src->name, header->source_name);
-  etcpal_timer_start(&src->packet_timer, SOURCE_LOSS_TIMEOUT);
-  src->seq = seq;
-  src->terminated = false;
-  src->dmx_received_since_last_tick = true;
-
-#if SACN_ETC_PRIORITY_EXTENSION
-  if (receiver->sampling)
-  {
-    if (header->start_code == SACN_STARTCODE_PRIORITY)
-    {
-      // Need to wait for DMX - ignore PAP packets until we've seen at least one DMX packet.
-      *notify = false;
-      src->recv_state = kRecvStateWaitingForDmx;
-      etcpal_timer_start(&src->pap_timer, SOURCE_LOSS_TIMEOUT);
-    }
-    else
-    {
-      // If we are in the sampling period, the wait period for PAP is not necessary.
-      src->recv_state = kRecvStateHaveDmxOnly;
-    }
-  }
-  else
-  {
-    // Even if this is a priority packet, we want to make sure that DMX packets are also being
-    // sent before notifying.
-    if (header->start_code == SACN_STARTCODE_PRIORITY)
-      src->recv_state = kRecvStateWaitingForDmx;
-    else
-      src->recv_state = kRecvStateWaitingForPap;
-    *notify = false;
-    etcpal_timer_start(&src->pap_timer, WAIT_FOR_PRIORITY);
-  }
-#endif
-
-  etcpal_rbtree_insert(&receiver->sources, src);
-  *new_source = src;
-
-  if (SACN_CAN_LOG(ETCPAL_LOG_DEBUG))
-  {
-    char cid_str[ETCPAL_UUID_STRING_BYTES];
-    etcpal_uuid_to_string(sender_cid, cid_str);
-    SACN_LOG_DEBUG("Tracking new source %s (%s) with initial start code 0x%02x", header->source_name, cid_str,
-                   header->start_code);
-  }
-}
-
-/*
- * Function that implements sACN's sequence numbering algorithm.
- *
- * [in] new_seq The sequence number accompanying the new packet.
- * [in] old_seq The most recent previous sequence number that was received.
- * Returns whether this packet is in sequence and should be processed.
- */
-bool check_sequence(int8_t new_seq, int8_t old_seq)
-{
-  int8_t seqnum_cmp = new_seq - old_seq;
-  return (seqnum_cmp > 0 || seqnum_cmp <= -20);
-}
-
-void deliver_receive_callbacks(const EtcPalSockAddr* from_addr, const EtcPalUuid* sender_cid,
-                               const SacnHeaderData* header, SourceLimitExceededNotification* source_limit_exceeded,
-                               SourcePapLostNotification* source_pap_lost, UniverseDataNotification* universe_data)
-{
-#if !SACN_LOGGING_ENABLED
-  ETCPAL_UNUSED_ARG(header);
-#endif
-
-  if (source_limit_exceeded->handle != SACN_RECEIVER_INVALID)
-  {
-    if (SACN_CAN_LOG(ETCPAL_LOG_WARNING))
-    {
-      char cid_str[ETCPAL_UUID_STRING_BYTES];
-      etcpal_uuid_to_string(sender_cid, cid_str);
-      SACN_LOG_WARNING(
-          "No room to track new sACN source %s (%s) on universe %u. This message will only be logged once each "
-          "time the maximum number of sources is exceeded.",
-          header->source_name, cid_str, header->universe_id);
-    }
-
-    if (source_limit_exceeded->callback)
-      source_limit_exceeded->callback(source_limit_exceeded->handle, source_limit_exceeded->universe,
-                                      source_limit_exceeded->context);
-  }
-
-  if (source_pap_lost->handle != SACN_RECEIVER_INVALID && source_pap_lost->callback)
-  {
-    source_pap_lost->callback(source_pap_lost->handle, source_pap_lost->universe, &source_pap_lost->source,
-                              source_pap_lost->context);
-  }
-
-  if (universe_data->handle != SACN_RECEIVER_INVALID && universe_data->callback)
-  {
-    universe_data->callback(universe_data->handle, from_addr, &universe_data->header, universe_data->pdata,
-                            universe_data->is_sampling, universe_data->context);
-  }
-}
-
-/**************************************************************************************************
- * Internal helpers for processing periodic timeout functionality
- *************************************************************************************************/
-
-/*
- * Handle periodic sACN Receive timeout functionality.
- * [in] recv_thread_context Context data for this thread.
- */
-void process_receivers(SacnRecvThreadContext* recv_thread_context)
-{
-  SamplingStartedNotification* sampling_started = NULL;
-  size_t num_sampling_started = 0;
-  SamplingEndedNotification* sampling_ended = NULL;
-  size_t num_sampling_ended = 0;
-  SourcesLostNotification* sources_lost = NULL;
-  size_t num_sources_lost = 0;
-
-  if (sacn_lock())
-  {
-    size_t num_receivers = recv_thread_context->num_receivers;
-
-    sampling_started = get_sampling_started_buffer(recv_thread_context->thread_id, num_receivers);
-    sampling_ended = get_sampling_ended_buffer(recv_thread_context->thread_id, num_receivers);
-    sources_lost = get_sources_lost_buffer(recv_thread_context->thread_id, num_receivers);
-    if (!sampling_started || !sampling_ended || !sources_lost)
-    {
-      sacn_unlock();
-      SACN_LOG_ERR("Could not allocate memory to track state data for sACN receivers!");
-      return;
-    }
-
-    for (SacnReceiver* receiver = recv_thread_context->receivers; receiver; receiver = receiver->next)
-    {
-      // Check the sample period
-      if (!receiver->notified_sampling_started)
-      {
-        receiver->notified_sampling_started = true;
-        sampling_started[num_sampling_started].callback = receiver->callbacks.sampling_period_started;
-        sampling_started[num_sampling_started].context = receiver->callbacks.context;
-        sampling_started[num_sampling_started].handle = receiver->keys.handle;
-        sampling_started[num_sampling_started].universe = receiver->keys.universe;
-        ++num_sampling_started;
-      }
-
-      if (receiver->sampling && etcpal_timer_is_expired(&receiver->sample_timer))
-      {
-        receiver->sampling = false;
-        sampling_ended[num_sampling_ended].callback = receiver->callbacks.sampling_period_ended;
-        sampling_ended[num_sampling_ended].context = receiver->callbacks.context;
-        sampling_ended[num_sampling_ended].handle = receiver->keys.handle;
-        sampling_ended[num_sampling_ended].universe = receiver->keys.universe;
-        ++num_sampling_ended;
-      }
-
-      process_receiver_sources(recv_thread_context->thread_id, receiver, &sources_lost[num_sources_lost++]);
-    }
-
-    sacn_unlock();
-  }
-
-  PeriodicCallbacks periodic_callbacks;
-  periodic_callbacks.sources_lost_arr = sources_lost;
-  periodic_callbacks.num_sources_lost = num_sources_lost;
-  periodic_callbacks.sampling_started_arr = sampling_started;
-  periodic_callbacks.num_sampling_started = num_sampling_started;
-  periodic_callbacks.sampling_ended_arr = sampling_ended;
-  periodic_callbacks.num_sampling_ended = num_sampling_ended;
-
-  deliver_periodic_callbacks(&periodic_callbacks);
-}
-
-void process_receiver_sources(sacn_thread_id_t thread_id, SacnReceiver* receiver, SourcesLostNotification* sources_lost)
-{
-  SacnSourceStatusLists* status_lists = get_status_lists(thread_id);
-  SacnTrackedSource** to_erase = get_to_erase_buffer(thread_id, etcpal_rbtree_size(&receiver->sources));
-
-  if (!status_lists || !to_erase)
-  {
-    SACN_LOG_ERR("Couldn't allocate memory to process sACN receiver for universe %u!", receiver->keys.universe);
-    return;
-  }
-  size_t num_to_erase = 0;
-
-  // And iterate through the sources on each universe
-  EtcPalRbIter src_it;
-  etcpal_rbiter_init(&src_it);
-  SacnTrackedSource* src = etcpal_rbiter_first(&src_it, &receiver->sources);
-  while (src)
-  {
-    if (!check_source_timeouts(src, status_lists))
-    {
-      to_erase[num_to_erase++] = src;
-      if (SACN_CAN_LOG(ETCPAL_LOG_DEBUG))
-      {
-        char cid_str[ETCPAL_UUID_STRING_BYTES];
-        etcpal_uuid_to_string(&src->cid, cid_str);
-        SACN_LOG_DEBUG("Removing internally tracked source %s", cid_str);
-      }
-    }
-
-    src = etcpal_rbiter_next(&src_it);
-  }
-
-  mark_sources_offline(status_lists->offline, status_lists->num_offline, status_lists->unknown,
-                       status_lists->num_unknown, &receiver->term_sets, receiver_state.expired_wait);
-  mark_sources_online(status_lists->online, status_lists->num_online, receiver->term_sets);
-  get_expired_sources(&receiver->term_sets, sources_lost);
-
-  for (size_t i = 0; i < num_to_erase; ++i)
-  {
-    etcpal_rbtree_remove(&receiver->sources, to_erase[i]);
-    FREE_TRACKED_SOURCE(to_erase[i]);
-  }
-
-  if (sources_lost->num_lost_sources > 0)
-  {
-    sources_lost->callback = receiver->callbacks.sources_lost;
-    sources_lost->context = receiver->callbacks.context;
-    sources_lost->handle = receiver->keys.handle;
-    sources_lost->universe = receiver->keys.universe;
-    for (size_t i = 0; i < sources_lost->num_lost_sources; ++i)
-    {
-      etcpal_rbtree_remove_with_cb(&receiver->sources, &sources_lost->lost_sources[i].cid, source_tree_dealloc);
-    }
-  }
-}
-
-/*
- * Check the various packet timeouts of a given source and add it to status lists if necessary.
- *
- * [in,out] src Source to be checked - some state is updated.
- * [out] status_lists Set of status lists to which to add the source based on its state.
- * Returns false if the source timed out while in a waiting state and should be removed immediately.
- */
-bool check_source_timeouts(SacnTrackedSource* src, SacnSourceStatusLists* status_lists)
-{
-  bool res = true;
-
-#if SACN_ETC_PRIORITY_EXTENSION
-
-  switch (src->recv_state)
-  {
-    case kRecvStateWaitingForDmx:
-      if (etcpal_timer_is_expired(&src->pap_timer))
-        res = false;
-      break;
-    case kRecvStateWaitingForPap:
-      if (etcpal_timer_is_expired(&src->packet_timer))
-        res = false;
-      break;
-    case kRecvStateHaveDmxOnly:
-    case kRecvStateHaveDmxAndPap:
-      update_source_status(src, status_lists);
-      break;
-    default:
-      break;
-  }
-
-#else  // SACN_ETC_PRIORITY_EXTENSION
-
-  update_source_status(src, status_lists);
-
-#endif  // SACN_ETC_PRIORITY_EXTENSION
-
-  return res;
-}
-
-void update_source_status(SacnTrackedSource* src, SacnSourceStatusLists* status_lists)
-{
-  if (etcpal_timer_is_expired(&src->packet_timer))
-  {
-    if (!add_offline_source(status_lists, &src->cid, src->name, src->terminated) && SACN_CAN_LOG(ETCPAL_LOG_ERR))
-    {
-      char cid_str[ETCPAL_UUID_BYTES];
-      etcpal_uuid_to_string(&src->cid, cid_str);
-      SACN_LOG_ERR(
-          "Couldn't allocate memory to add offline source %s to status list. This could be a bug or resource "
-          "exhaustion issue.",
-          cid_str);
-    }
-  }
-  else if (src->dmx_received_since_last_tick)
-  {
-    if (!add_online_source(status_lists, &src->cid, src->name) && SACN_CAN_LOG(ETCPAL_LOG_ERR))
-    {
-      char cid_str[ETCPAL_UUID_BYTES];
-      etcpal_uuid_to_string(&src->cid, cid_str);
-      SACN_LOG_ERR(
-          "Couldn't allocate memory to add online source %s to status list. This could be a bug or resource "
-          "exhaustion issue.",
-          cid_str);
-    }
-    src->dmx_received_since_last_tick = false;
-  }
-  else
-  {
-    if (!add_unknown_source(status_lists, &src->cid, src->name) && SACN_CAN_LOG(ETCPAL_LOG_ERR))
-    {
-      char cid_str[ETCPAL_UUID_BYTES];
-      etcpal_uuid_to_string(&src->cid, cid_str);
-      SACN_LOG_ERR(
-          "Couldn't allocate memory to add undetermined source %s to status list. This could be a bug or resource "
-          "exhaustion issue.",
-          cid_str);
-    }
-  }
-}
-
-void deliver_periodic_callbacks(const PeriodicCallbacks* periodic_callbacks)
-{
-  for (const SamplingStartedNotification* notif = periodic_callbacks->sampling_started_arr;
-       notif < periodic_callbacks->sampling_started_arr + periodic_callbacks->num_sampling_started; ++notif)
-  {
-    if (notif->callback)
-      notif->callback(notif->handle, notif->universe, notif->context);
-  }
-
-  for (const SamplingEndedNotification* notif = periodic_callbacks->sampling_ended_arr;
-       notif < periodic_callbacks->sampling_ended_arr + periodic_callbacks->num_sampling_ended; ++notif)
-  {
-    if (notif->callback)
-      notif->callback(notif->handle, notif->universe, notif->context);
-  }
-
-  for (const SourcesLostNotification* notif = periodic_callbacks->sources_lost_arr;
-       notif < periodic_callbacks->sources_lost_arr + periodic_callbacks->num_sources_lost; ++notif)
-  {
-    if (notif->callback)
-      notif->callback(notif->handle, notif->universe, notif->lost_sources, notif->num_lost_sources, notif->context);
-  }
-}
-
-/**************************************************************************************************
- * Internal helpers for managing the trees that track receivers and sources
- *************************************************************************************************/
-
-int receiver_compare(const EtcPalRbTree* tree, const void* value_a, const void* value_b)
-{
-  ETCPAL_UNUSED_ARG(tree);
-
-  const SacnReceiver* a = (const SacnReceiver*)value_a;
-  const SacnReceiver* b = (const SacnReceiver*)value_b;
-  return (a->keys.handle > b->keys.handle) - (a->keys.handle < b->keys.handle);
-}
-
-int receiver_compare_by_universe(const EtcPalRbTree* tree, const void* value_a, const void* value_b)
-{
-  ETCPAL_UNUSED_ARG(tree);
-
-  const SacnReceiver* a = (const SacnReceiver*)value_a;
-  const SacnReceiver* b = (const SacnReceiver*)value_b;
-  return (a->keys.universe > b->keys.universe) - (a->keys.universe < b->keys.universe);
-}
-
-int tracked_source_compare(const EtcPalRbTree* tree, const void* value_a, const void* value_b)
-{
-  ETCPAL_UNUSED_ARG(tree);
-
-  const SacnTrackedSource* a = (const SacnTrackedSource*)value_a;
-  const SacnTrackedSource* b = (const SacnTrackedSource*)value_b;
-  return ETCPAL_UUID_CMP(&a->cid, &b->cid);
-}
-
-EtcPalRbNode* node_alloc(void)
-{
-#if SACN_DYNAMIC_MEM
-  return (EtcPalRbNode*)malloc(sizeof(EtcPalRbNode));
-#else
-  return etcpal_mempool_alloc(sacnrecv_rb_nodes);
-#endif
-}
-
-void node_dealloc(EtcPalRbNode* node)
-{
-#if SACN_DYNAMIC_MEM
-  free(node);
-#else
-  etcpal_mempool_free(sacnrecv_rb_nodes, node);
-#endif
-}
-
-/* Helper function for clearing an EtcPalRbTree containing sources. */
-static void source_tree_dealloc(const EtcPalRbTree* self, EtcPalRbNode* node)
-{
-  ETCPAL_UNUSED_ARG(self);
-  FREE_TRACKED_SOURCE(node->value);
-  node_dealloc(node);
-}
-
-/* Helper function for clearing an EtcPalRbTree containing SacnReceivers. */
-static void universe_tree_dealloc(const EtcPalRbTree* self, EtcPalRbNode* node)
-{
-  ETCPAL_UNUSED_ARG(self);
-
-  SacnReceiver* receiver = (SacnReceiver*)node->value;
-  etcpal_rbtree_clear_with_cb(&receiver->sources, source_tree_dealloc);
-  remove_receiver_sockets(receiver, kCloseSocketNow);
-  FREE_RECEIVER(receiver);
-  node_dealloc(node);
-}
-
-bool receiver_handle_in_use(int handle_val, void* cookie)
-{
-  ETCPAL_UNUSED_ARG(cookie);
-
-  return (etcpal_rbtree_find(&receiver_state.receivers, &handle_val) != NULL);
 }
